Optimize coding threads (#3)
* Opimized ish...

+++ TimeTilTheEnd/Holiday.cs
@@ -10,7 +10,7 @@ namespace TimeTilTheEnd
     class Holiday
     {
         /// <summary>
-        /// create 2 dates
+        /// create 2 dates, start day and end day of a date
         /// </summary>
         /// <param name="startyear"></param>
         /// <param name="startmonth"></param>
@@ -19,7 +19,7 @@ namespace TimeTilTheEnd
         /// <param name="endmonth"></param>
         /// <param name="endday"></param>
         /// <returns></returns>
-        public DateTime[] Holidaay(int startyear, int startmonth, int startday, int endyear, int endmonth, int endday)
+        public DateTime[] Holidaay(int startyear, byte startmonth, byte startday, int endyear, byte endmonth, byte endday)
         {
             DateTime startdate = new DateTime(startyear, startmonth, startday);
             DateTime enddate = new DateTime(endyear, endmonth, endday);
